rename src files to fit origin/cross-origin paradigm
