+++ packages/driver/src/cross-origin/communicator.ts
