+++ packages/driver/src/cross-origin/events/errors.ts
