+++ packages/driver/src/cross-origin/events/logs.ts
