+++ packages/driver/src/cross-origin/events/misc.ts
