+++ packages/driver/src/cross-origin/events/screenshots.ts
