+++ packages/driver/src/cross-origin/events/spec_window.ts
