+++ packages/driver/src/cy/origin/validator.ts
