+++ packages/driver/src/cypress/commands.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash'
 import { allCommands } from '../cy/commands'
 import { addCommand as addNetstubbingCommand } from '../cy/net-stubbing'
-import { addCommands as addCrossOriginCommands } from '../cy/multi-domain'
+import { addCommands as addCrossOriginCommands } from '../cy/origin'
 import $errUtils from './error_utils'
 import $stackUtils from './stack_utils'
