+++ packages/driver/types/internal-types.d.ts
@@ -66,8 +66,8 @@ declare namespace Cypress {
     state: State
     events: Events
     emit: (event: string, payload?: any) => void
-    primaryOriginCommunicator: import('../src/multi-domain/communicator').PrimaryOriginCommunicator
-    specBridgeCommunicator: import('../src/multi-domain/communicator').SpecBridgeCommunicator
+    primaryOriginCommunicator: import('../src/cross-origin/communicator').PrimaryOriginCommunicator
+    specBridgeCommunicator: import('../src/cross-origin/communicator').SpecBridgeCommunicator
     mocha: $Mocha
     configure: (config: Cypress.ObjectLike) => void
     isCrossOriginSpecBridge: boolean
