+++ packages/runner/src/cross-origin.js
@@ -1,2 +1,2 @@
 // this is the entry point for the cross-origin version of the driver
-import '@packages/driver/src/multi-domain/cypress'
+import '@packages/driver/src/cross-origin/cypress'
